battery-status/permissions-policy/battery-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that battery is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissions-policy-features">
<link rel="help" href="https://w3c.github.io/battery/#permissions-policy-integration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('battery', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise battery.');
</script>
